requirements.txt
Flask==2.3.2
flask-swagger==0.2.14
flask-swagger-ui==4.11.1
pymongo==4.4.0
flask_cors==4.0.0
paddleocr==2.6.1.3
pdf2image==1.16.3
python-dotenv==1.0.0
langchain==0.0.219
boto3==1.26.164
PyPDF2==3.0.1
paddlepaddle==2.5.0
openai==0.27.8
tiktoken==0.4.0
faiss-cpu==1.7.4